+++ src/parser/inline.test.ts
@@ -11,17 +11,17 @@ describe('Unit: parser/inline', () => {
     });
 
     it('ab', () => {
-      assert.deepStrictEqual(inspect(parser('a`b`<s>')), [['a', '<code>b</code>', '&lt;', 's', '&gt;'], '']);
+      assert.deepStrictEqual(inspect(parser('a`b`<q>')), [['a', '<code>b</code>', '&lt;', 'q', '&gt;'], '']);
     });
 
     it('nest', () => {
-      assert.deepStrictEqual(inspect(parser('~~<var>`a`</var>~~')), [['<s><var><code>a</code></var></s>'], '']);
-      assert.deepStrictEqual(inspect(parser('<var>~~<dfn>a</dfn>~~</var>')), [['<var><s><dfn>a</dfn></s></var>'], '']);
+      assert.deepStrictEqual(inspect(parser('~~<var>`a`</var>~~')), [['<del><var><code>a</code></var></del>'], '']);
+      assert.deepStrictEqual(inspect(parser('<var>~~<dfn>a</dfn>~~</var>')), [['<var><del><dfn>a</dfn></del></var>'], '']);
     });
 
     it('flip', () => {
-      assert.deepStrictEqual(inspect(parser('**~~*<a>*~~**')), [['<strong><s><em>&lt;a&gt;</em></s></strong>'], '']);
-      assert.deepStrictEqual(inspect(parser('*~~**<a>**~~*')), [['<em><s><strong>&lt;a&gt;</strong></s></em>'], '']);
+      assert.deepStrictEqual(inspect(parser('**~~*<a>*~~**')), [['<strong><del><em>&lt;a&gt;</em></del></strong>'], '']);
+      assert.deepStrictEqual(inspect(parser('*~~**<a>**~~*')), [['<em><del><strong>&lt;a&gt;</strong></del></em>'], '']);
     });
 
   });