+++ src/parser/inline.ts
@@ -1,6 +1,6 @@
 ﻿import { MarkdownParser } from '../../markdown.d';
 import { combine } from '../combinator/combine';
-import { strike } from './inline/strike';
+import { deletion } from './inline/deletion';
 import { strong } from './inline/strong';
 import { emphasis } from './inline/emphasis';
 import { superscript } from './inline/superscript';
@@ -13,7 +13,7 @@ import { html } from './inline/html';
 import { text } from './inline/text';
 
 export import InlineParser = MarkdownParser.InlineParser;
-export import StrikeParser = InlineParser.StrikeParser;
+export import DeletionParser = InlineParser.DeletionParser;
 export import StrongParser = InlineParser.StrongParser;
 export import EmphasisParser = InlineParser.EmphasisParser;
 export import SuperScriptParser = InlineParser.SuperScriptParser;
@@ -27,7 +27,7 @@ export import TextParser = InlineParser.TextParser;
 export import PlainTextParser = InlineParser.PlainTextParser;
 
 export const inline: InlineParser = combine<[
-  StrikeParser,
+  DeletionParser,
   StrongParser,
   EmphasisParser,
   SuperScriptParser,
@@ -39,7 +39,7 @@ export const inline: InlineParser = combine<[
   HTMLParser,
   TextParser
 ], HTMLElement | Text>([
-  strike,
+  deletion,
   strong,
   emphasis,
   superscript,